[config] Refact the config task to use a template
Previously we were using lineinfile module to configure
Horizon. this method is not very easy and optimal because
the user has to know the option name.

Now we are using a template in templates/etc/openstack-dashboard/
local_settings.py.j2

We also add the possibility to use horizon_config_extras to configure
additional values not in the template.